FInally fixed all bugs with pathfinding, balloons work on every possible test thrown

+++ BTD/EnemyBase.pde
@@ -1,5 +1,5 @@
 public class EnemyBase extends Tower{
-  public EnemyBase(int x, int y, float maxHealth){
+  public EnemyBase(int x, int y, int maxHealth){
       super(x,y,maxHealth); 
   }
   public void display(){
